+++ TechnicalResearch/jetcobot/app/model/posture_model.py
@@ -4,10 +4,10 @@ from pydantic import BaseModel
 class PosturesData(BaseModel):
     pos_id: str
     pos_name: str
-    j1: int
-    j2: int
-    j3: int
-    j4: int
-    j5: int
-    j6: int
+    j1: float
+    j2: float
+    j3: float
+    j4: float
+    j5: float
+    j6: float
     gap: int
